--- js/renderPhoto.js
@@ -1,20 +1,0 @@
-const renderPictures = (posts) => {
-  const pictureTemplate = document.querySelector('#picture')
-    .content
-    .querySelector('.picture');
-  const pictureListElement = document.querySelector('.pictures');
-  const listFragment = document.createDocumentFragment();
-
-  posts.forEach((post) => {
-    const pictureElement = pictureTemplate.cloneNode(true);
-    pictureElement.dataset.id = post.id;
-    pictureElement.querySelector('.picture__img').src = post.url;
-    pictureElement.querySelector('.picture__comments').textContent = post.comments.length;
-    pictureElement.querySelector('.picture__likes').textContent = post.likes;
-    listFragment.appendChild(pictureElement);
-  });
-
-  pictureListElement.appendChild(listFragment);
-};
-
-export {renderPictures};
